python/requirements.txt
pandas
python-dotenv
git+https://github.com/robinhood-unofficial/pyrh.git#egg=pyrh